+++ src/main/kotlin/com/example/demo/controller/dto/InsertRequest.kt
@@ -1,0 +1,3 @@
+package com.example.demo.controller.dto
+
+data class InsertRequest(var helloWorld: String)
