+++ src/main/kotlin/com/example/demo/infrastructure/mapper/SampleMapper.kt
@@ -1,18 +1,18 @@
-package com.example.demo.mapper
+package com.example.demo.infrastructure.mapper
 
 import org.apache.ibatis.annotations.Mapper
-import com.example.demo.model.TodoModel
+import com.example.demo.model.SampleModel
 import org.apache.ibatis.annotations.Options
 import org.springframework.stereotype.Component
 
 @Mapper
 @Component
-interface TodoMapper {
+interface SampleMapper {
     // 抽象メソッドで定義する
-    fun find(): List<TodoModel>
+    fun find(): List<SampleModel>
 
-    //@Options(useGeneratedKeys = true, keyProperty = "id")
-    fun insert(title:String)
+    @Options(useGeneratedKeys = true, keyProperty = "id")
+    fun insert(hello_world:String)
 //    fun insert(model: SampleModel)
 
     fun update(id: Int, hello_world: String)
