+++ src/main/kotlin/com/example/demo/infrastructure/mapper/TodoMapper.kt
@@ -1,18 +1,17 @@
-package com.example.demo.mapper
+package com.example.demo.infrastructure.mapper
 
 import org.apache.ibatis.annotations.Mapper
-import com.example.demo.model.SampleModel
-import org.apache.ibatis.annotations.Options
+import com.example.demo.model.TodoModel
 import org.springframework.stereotype.Component
 
 @Mapper
 @Component
-interface SampleMapper {
+interface TodoMapper {
     // 抽象メソッドで定義する
-    fun find(): List<SampleModel>
+    fun find(): List<TodoModel>
 
-    @Options(useGeneratedKeys = true, keyProperty = "id")
-    fun insert(hello_world:String)
+    //@Options(useGeneratedKeys = true, keyProperty = "id")
+    fun insert(title:String)
 //    fun insert(model: SampleModel)
 
     fun update(id: Int, hello_world: String)
