+++ src/main/kotlin/com/example/demo/model/SampleModel.kt
@@ -2,7 +2,6 @@ package com.example.demo.model
 
 class SampleModel(var id: Int, var helloWorld: String)
 
-data class insertRequest(var helloWorld: String)
 
 //class SampleModel{
 //    private var id:Int = 0
