+++ src/main/resources/mybatis/mapper/SampleMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.demo.mapper.SampleMapper">
+<mapper namespace="com.example.demo.infrastructure.mapper.SampleMapper">
 
     <select id="find" resultType="SampleModel">
         SELECT
