+++ src/main/resources/mybatis/mapper/TodoMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.demo.mapper.TodoMapper">
+<mapper namespace="com.example.demo.infrastructure.mapper.TodoMapper">
 
     <select id="find" resultType="TodoModel">
         SELECT
